+++ lib/options/rest_api_client_request_options.dart
@@ -1,0 +1,18 @@
+import 'package:dio/dio.dart';
+
+class RestApiClientRequestOptions {
+  Map<String, dynamic>? headers;
+  String? contentType;
+
+  RestApiClientRequestOptions({
+    this.headers,
+    this.contentType,
+  });
+
+  Options toOptions() {
+    return Options(
+      headers: headers,
+      contentType: contentType,
+    );
+  }
+}